Replace  and  with versions using new class
