+++ scripts/plotting/global_latlon_map.py
@@ -8,6 +8,8 @@ global_latlon_map(adfobj)
 my_formatwarning(msg, *args, **kwargs)
     format warning messages
     (private method)
+plot_file_op
+    Check on status of output plot file.
 """
 #Import standard modules:
 from pathlib import Path
@@ -15,13 +17,15 @@ import numpy as np
 import xarray as xr
 import warnings  # use to warn user about missing files.
 
+import plotting_functions as pf
+
 #Format warning messages:
 def my_formatwarning(msg, *args, **kwargs):
     """Issue `msg` as warning."""
     return str(msg) + '\n'
 warnings.formatwarning = my_formatwarning
 
-def global_latlon_map(adfobj):
+def global_latlon_map_B(adfobj):
     """
     This script/function is designed to generate global
     2-D lat/lon maps of model fields with continental overlays.
@@ -37,32 +41,22 @@ def global_latlon_map(adfobj):
 
     Notes
     -----
-    This function imports `pandas` and `plotting_functions`
 
     It uses the AdfDiag object's methods to get necessary information.
-    Specificially:
+    This version passes the AdfDiag object to instatiate an AdfData object.
+    Explicitly accesses:
     adfobj.diag_var_list
         List of variables
-    adfobj.get_basic_info
-        Regrid data path, checks `compare_obs`, checks `redo_plot`, checks `plot_press_levels`
     adfobj.plot_location
         output plot path
-    adfobj.get_cam_info
-        Get `cam_case_name` and `case_nickname`
     adfobj.climo_yrs
         start and end climo years of the case(s), `syears` & `eyears`
         start and end climo years of the reference, `syear_baseline` & `eyear_baseline`
-    adfobj.var_obs_dict
-        reference data (conditional)
-    adfobj.get_baseline_info
-        get reference case, `cam_case_name`
     adfobj.variable_defaults 
         dict of variable-specific plot preferences
     adfobj.read_config_var
         dict of basic info, `diag_basic_info`
         Then use to check `plot_type`
-    adfobj.compare_obs
-        Used to set data path
     adfobj.debug_log
         Issues debug message
     adfobj.add_website_data
@@ -82,20 +76,6 @@ def global_latlon_map(adfobj):
         Checks on pressure level dimension
     """
 
-    #Import necessary modules:
-    #------------------------
-    import pandas as pd
-
-    #CAM diagnostic plotting functions:
-    import plotting_functions as pf
-    #-------------------------
-
-    # Steps:
-    # - load regridded climo files for model and obs
-    # - calculate all-time and seasonal fields (from individual months)
-    # - Take difference, calculate statistics
-    # - make plot
-
     #Notify user that script has started:
     print("\n  Generating lat/lon maps...")
 
@@ -103,49 +83,17 @@ def global_latlon_map(adfobj):
     # Use ADF api to get all necessary information
     #
     var_list = adfobj.diag_var_list
-    model_rgrid_loc = adfobj.get_basic_info("cam_regrid_loc", required=True)
-
     #Special ADF variable which contains the output paths for
     #all generated plots and tables for each case:
     plot_locations = adfobj.plot_location
-
-    #CAM simulation variables (this is always assumed to be a list):
-    case_names = adfobj.get_cam_info("cam_case_name", required=True)
 
     #Grab case years
     syear_cases = adfobj.climo_yrs["syears"]
     eyear_cases = adfobj.climo_yrs["eyears"]
 
-    # CAUTION:
-    # "data" here refers to either obs or a baseline simulation,
-    # Until those are both treated the same (via intake-esm or similar)
-    # we will do a simple check and switch options as needed:
-    if adfobj.get_basic_info("compare_obs"):
-        #Set obs call for observation details for plot titles
-        obs = True
-
-        #Extract variable-obs dictionary:
-        var_obs_dict = adfobj.var_obs_dict
-
-        #If dictionary is empty, then  there are no observations to regrid to,
-        #so quit here:
-        if not var_obs_dict:
-            print("No observations found to plot against, so no lat/lon maps will be generated.")
-            return
-    else:
-        obs = False
-        data_name = adfobj.get_baseline_info("cam_case_name", required=True) # does not get used, is just here as a placemarker
-        data_list = [data_name] # gets used as just the name to search for climo files HAS TO BE LIST
-        data_loc  = model_rgrid_loc #Just use the re-gridded model data path
-    #End if
-
     #Grab baseline years (which may be empty strings if using Obs):
     syear_baseline = adfobj.climo_yrs["syear_baseline"]
     eyear_baseline = adfobj.climo_yrs["eyear_baseline"]
-
-    #Grab all case nickname(s)
-    test_nicknames = adfobj.case_nicknames["test_nicknames"]
-    base_nickname = adfobj.case_nicknames["base_nickname"]
 
     res = adfobj.variable_defaults # will be dict of variable-specific plot preferences
     # or an empty dictionary if use_defaults was not specified in YAML.
@@ -162,19 +110,11 @@ def global_latlon_map(adfobj):
     print(f"\t NOTE: redo_plot is set to {redo_plot}")
     #-----------------------------------------
 
-    #Set data path variables:
-    #-----------------------
-    mclimo_rg_loc = Path(model_rgrid_loc)
-    if not adfobj.compare_obs:
-        dclimo_loc  = Path(data_loc)
-    #-----------------------
-
     #Determine if user wants to plot 3-D variables on
     #pressure levels:
     pres_levs = adfobj.get_basic_info("plot_press_levels")
 
-    #For now, let's always do seasonal weighting:
-    weight_season = True
+    weight_season = True  #always do seasonal weighting
 
     #Set seasonal ranges:
     seasons = {"ANN": np.arange(1,13,1),
@@ -183,29 +123,13 @@ def global_latlon_map(adfobj):
                "MAM": [3, 4, 5],
                "SON": [9, 10, 11]
                }
-
+    
     # probably want to do this one variable at a time:
     for var in var_list:
-
-        if adfobj.compare_obs:
-            #Check if obs exist for the variable:
-            if var in var_obs_dict:
-                #Note: In the future these may all be lists, but for
-                #now just convert the target_list.
-                #Extract target file:
-                dclimo_loc = var_obs_dict[var]["obs_file"]
-                #Extract target list (eventually will be a list, for now need to convert):
-                data_list = [var_obs_dict[var]["obs_name"]]
-                #Extract target variable name:
-                data_var = var_obs_dict[var]["obs_var"]
-            else:
-                dmsg = f"No obs found for variable `{var}`, lat/lon map plotting skipped."
-                adfobj.debug_log(dmsg)
-                continue
-        else:
-            #Set "data_var" for consistent use below:
-            data_var = var
-        #End if
+        if var not in adfobj.data.ref_var_nam:
+            dmsg = f"No reference data found for variable `{var}`, zonal mean plotting skipped."
+            adfobj.debug_log(dmsg)
+            continue        
 
         #Notify user of variable being plotted:
         print("\t - lat/lon maps for {}".format(var))
@@ -229,258 +153,202 @@ def global_latlon_map(adfobj):
         # otherwise defaults to 180
         vres['central_longitude'] = pf.get_central_longitude(adfobj)
 
-        #loop over different data sets to plot model against:
-        for data_src in data_list:
+        # load reference data (observational or baseline)
+        odata = adfobj.data.load_reference_da(var)
+        if odata is None:
+            continue
+        has_dims = pf.lat_lon_validate_dims(odata) # T iff dims are (lat,lon) -- can't plot unless we have both
+        if not has_dims:
+            print(f"\t = skipping global map for {var} as REFERENCE does not have both lat and lon")
+            continue
 
-            # load data (observational) commparison files (we should explore intake as an alternative to having this kind of repeated code):
-            if adfobj.compare_obs:
-                #For now, only grab one file (but convert to list for use below)
-                oclim_fils = [dclimo_loc]
-            else:
-                oclim_fils = sorted(dclimo_loc.glob(f"{data_src}_{var}_baseline.nc"))
+        #Loop over model cases:
+        for case_idx, case_name in enumerate(adfobj.data.case_names):
 
-            oclim_ds = pf.load_dataset(oclim_fils)
-            if oclim_ds is None:
-                print("WARNING: Did not find any oclim_fils. Will try to skip.")
-                print(f"INFO: Data Location, dclimo_loc is {dclimo_loc}")
-                print(f"INFO: The glob is: {data_src}_{var}_*.nc")
+            #Set case nickname:
+            case_nickname = adfobj.data.test_nicknames[case_idx]
+
+            #Set output plot location:
+            plot_loc = Path(plot_locations[case_idx])
+
+            #Check if plot output directory exists, and if not, then create it:
+            if not plot_loc.is_dir():
+                print("    {} not found, making new directory".format(plot_loc))
+                plot_loc.mkdir(parents=True)
+
+            #Load re-gridded model files:
+            mdata = adfobj.data.load_regrid_da(case_name, var)
+
+            #Skip this variable/case if the regridded climo file doesn't exist:
+            if mdata is None:
                 continue
-            #End if
 
-            #Loop over model cases:
-            for case_idx, case_name in enumerate(case_names):
-
-                #Set case nickname:
-                case_nickname = test_nicknames[case_idx]
-
-                #Set output plot location:
-                plot_loc = Path(plot_locations[case_idx])
-
-                #Check if plot output directory exists, and if not, then create it:
-                if not plot_loc.is_dir():
-                    print("    {} not found, making new directory".format(plot_loc))
-                    plot_loc.mkdir(parents=True)
-
-                #Load re-gridded model files:
-                mclim_fils = sorted(mclimo_rg_loc.glob(f"{data_src}_{case_name}_{var}_*.nc"))
-                mclim_ds = pf.load_dataset(mclim_fils)
-
-                #Skip this variable/case if the regridded climo file doesn't exist:
-                if mclim_ds is None:
-                    print("WARNING: Did not find any regridded climo files. Will try to skip.")
-                    print(f"INFO: Data Location, mclimo_rg_loc, is {mclimo_rg_loc}")
-                    print(f"INFO: The glob is: {data_src}_{case_name}_{var}_*.nc")
+            #Determine dimensions of variable:
+            has_dims_cam = pf.lat_lon_validate_dims(mdata) # T iff dims are (lat,lon) -- can't plot unless we have both
+            _, has_lev = pf.zm_validate_dims(mdata)    # has_lev T if lev in mdata
+            if not has_dims_cam:
+                print(f"\t = skipping global map for {var} for case {case_name} as it does not have both lat and lon")
+                continue
+            else: # i.e., has lat&lon
+                if pres_levs and (not has_lev):
+                    print(f"\t - skipping global map for {var} as it has more than lat/lon dims, but no pressure levels were provided")
                     continue
-                #End if
 
-                #Extract variable of interest
-                odata = oclim_ds[data_var].squeeze()  # squeeze in case of degenerate dimensions
-                mdata = mclim_ds[var].squeeze()
+            # Check output file. If file does not exist, proceed.
+            # If file exists:
+            #   if redo_plot is true: delete it now and make plot
+            #   if redo_plot is false: add to website and move on
+            doplot = {}
+            if not pres_levs:
+                for s in seasons:
+                    plot_name = plot_loc / f"{var}_{s}_LatLon_Mean.{plot_type}"
+                    doplot[plot_name] = plot_file_op(adfobj, plot_name, var, case_name, s, web_category, redo_plot, "LatLon")
+            else:
+                for pres in pres_levs:
+                    for s in seasons:
+                        plot_name = plot_loc / f"{var}_{pres}hpa_{s}_LatLon_Mean.{plot_type}"
+                        doplot[plot_name] = plot_file_op(adfobj, plot_name, f"{var}_{pres}hpa", case_name, s, web_category, redo_plot, "LatLon")
+            if all(value is None for value in doplot.values()):
+                print(f"All plots exist for {var}. Redo is {redo_plot}. Existing plots added to website data. Continue.")
+                continue
 
-                # APPLY UNITS TRANSFORMATION IF SPECIFIED:
-                # NOTE: looks like our climo files don't have all their metadata
-                mdata = mdata * vres.get("scale_factor",1) + vres.get("add_offset", 0)
-                # update units
-                mdata.attrs['units'] = vres.get("new_unit", mdata.attrs.get('units', 'none'))
+            #Create new dictionaries:
+            mseasons = {}
+            oseasons = {}
+            dseasons = {} # hold the differences
 
-                # Do the same for the baseline case if need be:
-                if not adfobj.compare_obs:
-                    odata = odata * vres.get("scale_factor",1) + vres.get("add_offset", 0)
-                    # update units
-                    odata.attrs['units'] = vres.get("new_unit", odata.attrs.get('units', 'none'))
-                # Or for observations:
-                else:
-                    odata = odata * vres.get("obs_scale_factor",1) + vres.get("obs_add_offset", 0)
-                   # Note: we are going to assume that the specification ensures the conversion makes the units the same. Doesn't make sense to add a different unit.
+            if not pres_levs:            
 
-                #Determine dimensions of variable:
-                has_dims = pf.lat_lon_validate_dims(odata)
-                if has_dims:
-                    #If observations/baseline CAM have the correct
-                    #dimensions, does the input CAM run have correct
-                    #dimensions as well?
-                    has_dims_cam = pf.lat_lon_validate_dims(mdata)
+                #Loop over season dictionary:
+                for s in seasons:
+                    plot_name = plot_loc / f"{var}_{s}_LatLon_Mean.{plot_type}"
+                    if doplot[plot_name] is None:
+                        continue
 
-                    #If both fields have the required dimensions, then
-                    #proceed with plotting:
-                    if has_dims_cam:
+                    if weight_season:
+                        mseasons[s] = pf.seasonal_mean(mdata, season=s, is_climo=True)
+                        oseasons[s] = pf.seasonal_mean(odata, season=s, is_climo=True)
+                    else:
+                        #Just average months as-is:
+                        mseasons[s] = mdata.sel(time=seasons[s]).mean(dim='time')
+                        oseasons[s] = odata.sel(time=seasons[s]).mean(dim='time')
+                    #End if
 
-                        #
-                        # Seasonal Averages
-                        # Note: xarray can do seasonal averaging,
-                        # but depends on having time accessor,
-                        # which these prototype climo files do not have.
-                        #
+                    # difference: each entry should be (lat, lon)
+                    dseasons[s] = mseasons[s] - oseasons[s]
 
-                        #Create new dictionaries:
-                        mseasons = {}
-                        oseasons = {}
-                        dseasons = {} # hold the differences
+                    pf.plot_map_and_save(plot_name, case_nickname, adfobj.data.ref_nickname,
+                                            [syear_cases[case_idx],eyear_cases[case_idx]],
+                                            [syear_baseline,eyear_baseline],
+                                            mseasons[s], oseasons[s], dseasons[s],
+                                            obs=adfobj.compare_obs, **vres)
 
-                        #Loop over season dictionary:
-                        for s in seasons:
-                            # time to make plot; here we'd probably loop over whatever plots we want for this variable
-                            # I'll just call this one "LatLon_Mean"  ... would this work as a pattern [operation]_[AxesDescription] ?
-                            plot_name = plot_loc / f"{var}_{s}_LatLon_Mean.{plot_type}"
+                    #Add plot to website (if enabled):
+                    adfobj.add_website_data(plot_name, var, case_name, category=web_category,
+                                            season=s, plot_type="LatLon")
 
-                            # Check redo_plot. If set to True: remove old plot, if it already exists:
-                            if (not redo_plot) and plot_name.is_file():
-                                #Add already-existing plot to website (if enabled):
-                                adfobj.debug_log(f"'{plot_name}' exists and clobber is false.")
-                                adfobj.add_website_data(plot_name, var, case_name, category=web_category,
-                                                        season=s, plot_type="LatLon")
+            else: # => pres_levs has values, & we already checked that lev is in mdata (has_lev)
 
-                                #Continue to next iteration:
-                                continue
-                            elif (redo_plot) and plot_name.is_file():
-                                plot_name.unlink()
+                for pres in pres_levs:
 
+                    #Check that the user-requested pressure level
+                    #exists in the model data, which should already
+                    #have been interpolated to the standard reference
+                    #pressure levels:
+                    if not (pres in mdata['lev']):
+                        print(f"plot_press_levels value '{pres}' not present in {var}, so skipping.")
+                        continue
 
-                            if weight_season:
-                                mseasons[s] = pf.seasonal_mean(mdata, season=s, is_climo=True)
-                                oseasons[s] = pf.seasonal_mean(odata, season=s, is_climo=True)
-                            else:
-                                #Just average months as-is:
-                                mseasons[s] = mdata.sel(time=seasons[s]).mean(dim='time')
-                                oseasons[s] = odata.sel(time=seasons[s]).mean(dim='time')
-                            #End if
+                    #Loop over seasons:
+                    for s in seasons:
+                        plot_name = plot_loc / f"{var}_{pres}hpa_{s}_LatLon_Mean.{plot_type}"
+                        if doplot[plot_name] is None:
+                            continue
 
-                            # difference: each entry should be (lat, lon)
-                            dseasons[s] = mseasons[s] - oseasons[s]
-
-
-                            #Create new plot:
-                            # NOTE: send vres as kwarg dictionary.  --> ONLY vres, not the full res
-                            # This relies on `plot_map_and_save` knowing how to deal with the options
-                            # currently knows how to handle:
-                            #   colormap, contour_levels, diff_colormap, diff_contour_levels, tiString, tiFontSize, mpl
-                            #   *Any other entries will be ignored.
-                            # NOTE: If we were doing all the plotting here, we could use whatever we want from the provided YAML file.
-
-                            pf.plot_map_and_save(plot_name, case_nickname, base_nickname,
-                                                 [syear_cases[case_idx],eyear_cases[case_idx]],
-                                                 [syear_baseline,eyear_baseline],
-                                                 mseasons[s], oseasons[s], dseasons[s],
-                                                 obs, **vres)
-
-                            #Add plot to website (if enabled):
-                            adfobj.add_website_data(plot_name, var, case_name, category=web_category,
-                                                    season=s, plot_type="LatLon")
-
-                    else: #mdata dimensions check
-                        print(f"\t - skipping lat/lon map for {var} as it doesn't have only lat/lon dims.")
-                    #End if (dimensions check)
-
-                elif pres_levs: #Is the user wanting to interpolate to a specific pressure level?
-
-                    #Check that case inputs have the correct dimensions (including "lev"):
-                    _, has_lev = pf.zm_validate_dims(mdata)
-
-                    if has_lev:
-
-                        #Calculate monthly weights (if applicable):
                         if weight_season:
-                            #Add date-stamp to time dimension:
-                            #Note: For now using made-up dates, but in the future
-                            #it might be good to extract this info from the files
-                            #themselves.
-                            timefix = pd.date_range(start='1/1/1980', end='12/1/1980', freq='MS')
-                            mdata['time']=timefix
-                            odata['time']=timefix
-
-                            #Calculate monthly weights based on number of days:
-                            month_length = mdata.time.dt.days_in_month
-                            weights = (month_length.groupby("time.season") / month_length.groupby("time.season").sum())
+                            mseasons[s] = pf.seasonal_mean(mdata, season=s, is_climo=True)
+                            oseasons[s] = pf.seasonal_mean(odata, season=s, is_climo=True)
+                        else:
+                            #Just average months as-is:
+                            mseasons[s] = mdata.sel(time=seasons[s]).mean(dim='time')
+                            oseasons[s] = odata.sel(time=seasons[s]).mean(dim='time')
                         #End if
 
-                        #Loop over pressure levels:
-                        for pres in pres_levs:
+                        # difference: each entry should be (lat, lon)
+                        dseasons[s] = mseasons[s] - oseasons[s]
 
-                            #Check that the user-requested pressure level
-                            #exists in the model data, which should already
-                            #have been interpolated to the standard reference
-                            #pressure levels:
-                            if not (pres in mclim_ds['lev']):
-                                #Move on to the next pressure level:
-                                print(f"plot_press_levels value '{pres}' not a standard reference pressure, so skipping.")
-                                continue
-                            #End if
+                        pf.plot_map_and_save(plot_name, case_nickname, data.ref_nickname,
+                                                [syear_cases[case_idx],eyear_cases[case_idx]],
+                                                [syear_baseline,eyear_baseline],
+                                                mseasons[s].sel(lev=pres), oseasons[s].sel(lev=pres), dseasons[s].sel(lev=pres),
+                                                obs=adfobj.compare_obs, **vres)
 
-                            #Create new dictionaries:
-                            mseasons = {}
-                            oseasons = {}
-                            dseasons = {}
-
-                            #Loop over seasons:
-                            for s in seasons:
-                                plot_name = plot_loc / f"{var}_{pres}hpa_{s}_LatLon_Mean.{plot_type}"
-
-                                # Check redo_plot. If set to True: remove old plot, if it already exists:
-                                redo_plot = adfobj.get_basic_info('redo_plot')
-                                if (not redo_plot) and plot_name.is_file():
-                                    #Add already-existing plot to website (if enabled):
-                                    adfobj.debug_log(f"'{plot_name}' exists and clobber is false.")
-                                    adfobj.add_website_data(plot_name, f"{var}_{pres}hpa", case_name, category=web_category,
-                                                            season=s, plot_type="LatLon")
-
-                                    #Continue to next iteration:
-                                    continue
-                                elif (redo_plot) and plot_name.is_file():
-                                    plot_name.unlink()
-
-                                #If requested, then calculate the monthly-weighted seasonal averages:
-                                if weight_season:
-                                    mseasons[s] = (pf.seasonal_mean(mdata, season=s, is_climo=True)).sel(lev=pres)
-                                    oseasons[s] = (pf.seasonal_mean(odata, season=s, is_climo=True)).sel(lev=pres)
-                                else:
-                                    #Just average months as-is:
-                                    mseasons[s] = mdata.sel(time=seasons[s], lev=pres).mean(dim='time')
-                                    oseasons[s] = odata.sel(time=seasons[s], lev=pres).mean(dim='time')
-                                #End if
-
-                                # difference: each entry should be (lat, lon)
-                                dseasons[s] = mseasons[s] - oseasons[s]
-
-                                # time to make plot; here we'd probably loop over whatever plots we want for this variable
-                                # I'll just call this one "LatLon_Mean"  ... would this work as a pattern [operation]_[AxesDescription] ?
-
-                                #Create new plot:
-                                # NOTE: send vres as kwarg dictionary.  --> ONLY vres, not the full res
-                                # This relies on `plot_map_and_save` knowing how to deal with the options
-                                # currently knows how to handle:
-                                #   colormap, contour_levels, diff_colormap, diff_contour_levels, tiString, tiFontSize, mpl
-                                #   *Any other entries will be ignored.
-                                # NOTE: If we were doing all the plotting here, we could use whatever we want from the provided YAML file.
-                                pf.plot_map_and_save(plot_name, case_nickname, base_nickname,
-                                                     [syear_cases[case_idx],eyear_cases[case_idx]],
-                                                     [syear_baseline,eyear_baseline],
-                                                     mseasons[s], oseasons[s], dseasons[s],
-                                                     obs, **vres)
-
-                                #Add plot to website (if enabled):
-                                adfobj.add_website_data(plot_name, f"{var}_{pres}hpa", case_name, category=web_category,
-                                                        season=s, plot_type="LatLon")
-
-                            #End for (seasons)
-                        #End for (pressure levels)
-
-                    else:
-                        print(f"\t - variable '{var}' has no vertical dimension but is not just time/lat/lon, so skipping.")
-                    #End if (has_lev)
-                else:
-                    print(f"\t - skipping polar map for {var} as it has more than lat/lon dims, but no pressure levels were provided")
-                #End if (dimensions check and plotting pressure levels)
-            #End for (case loop)
-        #End for (obs/baseline loop)
+                        #Add plot to website (if enabled):
+                        adfobj.add_website_data(plot_name, f"{var}_{pres}hpa", case_name, category=web_category,
+                                                season=s, plot_type="LatLon")
+                    #End for (seasons)
+                #End for (pressure levels)
+            #End if (plotting pressure levels)
+        #End for (case loop)
     #End for (variable loop)
 
     #Notify user that script has ended:
     print("  ...lat/lon maps have been generated successfully.")
 
-#########
-# Helpers
-#########
 
+def plot_file_op(adfobj, plot_name, var, case_name, season, web_category, redo_plot, plot_type):
+    """Check if output plot needs to be made or remade.
+    
+    Parameters
+    ----------
+    adfobj : AdfDiag
+        The diagnostics object that contains all the configuration information
+
+    plot_name : Path
+        path of the output plot
+
+    var : str
+        name of variable
+
+    case_name : str
+        case name
+    
+    season : str
+        season being plotted
+
+    web_category : str
+        the category for this variable
+
+    redo_plot : bool
+        whether to overwrite existing plot with this file name
+
+    plot_type : str
+        the file type for the output plot
+
+    Returns
+    -------
+    int, None
+        Returns 1 if existing file is removed or no existing file.
+        Returns None if file exists and redo_plot is False
+
+    Notes
+    -----
+    The long list of parameters is because add_website_data is called
+    when the file exists and will not be overwritten.
+    
+    """
+    # Check redo_plot. If set to True: remove old plot, if it already exists:
+    if plot_name.is_file():
+        if redo_plot:
+            plot_name.unlink()
+            return True
+        else:
+            #Add already-existing plot to website (if enabled):
+            adfobj.add_website_data(plot_name, var, case_name, category=web_category,
+                                    season=season, plot_type=plot_type)
+            return False  # False tells caller that file exists and not to overwrite
+    else:
+        return True
 
 ##############
 #END OF SCRIPT
